refactored the portfolio components for better organisation and readability
