+++ src/components/coinsPage/CoinSlider/CarouselTitle.tsx
@@ -1,8 +1,8 @@
 const CarouselTitle = ({ title }: { title: string }) => {
   return (
-    <h3 className="mb-8 font-grotesk text-sm text-[var(--clr-nav-text)]">
+    <h2 className="mb-8 font-grotesk text-sm text-[var(--clr-nav-text)]">
       {title}
-    </h3>
+    </h2>
   );
 };
