+++ src/components/convertorPage/ConvertorHeader/index.tsx
@@ -3,14 +3,14 @@
 import { useAppSelector } from "@/lib/hooks";
 import { getDate, getTime } from "@/utils/formatUtils";
 
-const ConvertorTitle = () => {
+const ConvertorHeader = () => {
   const { locale } = useAppSelector((state) => state.user);
   const date = getDate(undefined, { isShort: "short", locale: locale });
   const time = getTime(locale);
 
   return (
     <div className="mb-6 font-grotesk text-[var(--clr-nav-text)]">
-      <h3 className="mb-1 text-xl font-medium">Online currency convertor</h3>
+      <h2 className="mb-1 text-xl font-medium">Online currency convertor</h2>
       <span>
         {date} {time}
       </span>
@@ -18,4 +18,4 @@ const ConvertorTitle = () => {
   );
 };
 
-export default ConvertorTitle;
+export default ConvertorHeader;
